Backend/requirements.txt
absl-py==2.1.0
aiodataloader==0.4.0
aiosqlite==0.20.0
amqp==5.2.0
aniso8601==7.0.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
asyncio==3.4.3
beautifulsoup4==4.12.3
billiard==4.2.0
black==23.11.0
blinker==1.7.0
Brotli==1.1.0
celery==5.4.0
certifi==2018.11.29
cffi==1.16.0
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.7
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
databases==0.8.0
deepface==0.0.92
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.7
dj-stripe==2.8.4
Django==4.1.13
django-allauth==0.51.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-yasg==1.21.7
easyocr==1.7.1
email_validator==2.1.1
et-xmlfile==1.1.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.13.1
fire==0.6.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.2.0
gast==0.6.0
gdown==5.2.0
gevent==24.2.1
geventhttpclient==2.2.1
google-pasta==0.2.0
graphene==2.1.9
graphene-sqlalchemy==2.3.0
graphql-core==3.2.3
graphql-relay==2.0.1
graphql-server==3.0.0b7
greenlet==3.0.3
grpcio==1.65.1
grpcio-tools==1.65.1
gunicorn==21.2.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.3
idna==3.7
imageio==2.34.2
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.0
keras==3.4.1
kiwisolver==1.4.5
kombu==5.3.7
lazy_loader==0.4
libclang==18.1.1
locust==2.27.0
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.0.8
mtcnn==0.1.1
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
mysqlclient==2.2.4
namex==0.0.8
networkx==3.3
ninja==1.11.1.1
numpy==1.26.2
oauthlib==3.0.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pathspec==0.11.2
pbr==6.0.0
phonenumbers==8.13.27
pillow==10.4.0
pipenv==2023.11.15
platformdirs==4.1.0
promise==2.3
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==5.9.8
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pycparser==2.21
pydantic==2.7.3
pydantic_core==2.18.4
pygame==2.5.2
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
pyparsing==3.1.2
PySocks==1.7.1
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
python3-openid==3.1.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qiskit==1.1.0
redis==5.0.7
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.2.0
retina-face==0.0.17
rich==13.7.1
rustworkx==0.14.2
Rx==1.6.3
safetensors==0.4.2
scikit-image==0.24.0
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
setuptools==69.2.0
shapely==2.0.5
shellingham==1.5.4
singledispatch==3.7.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==1.4.52
sqlparse==0.5.1
starlette==0.37.2
stevedore==5.2.0
strawberry-graphql==0.234.0
stripe==3.5.0
symengine==0.11.0
sympy==1.12
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
tf_keras==2.17.0
threadpoolctl==3.4.0
tifffile==2024.7.24
tokenizers==0.15.2
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.2
transformers==4.38.2
typer==0.12.3
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.10.0
ultralytics==8.2.70
ultralytics-thop==2.0.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
vine==5.1.0
virtualenv==20.25.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.1
wheel==0.43.0
whitenoise==6.6.0
wrapt==1.16.0
zope.event==5.0
zope.interface==6.3
